+++ docs/changelog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2025-12-04
+- Fix the build job lookup algorithm in the integration test workflow.
+
 ## 2025-12-02
 - Add 5 more integration test secrets mapping slots.
 - Update the integration test workflow to run the build process concurrently with the setup phase of the integration tests.